reorganize nested-tabs-1 and screenshots
