+++ tests/playwright/sanity/modules/nested-tabs/nested-tabs-1.test.ts-snapshots/tab-title-with-html-linux.png
