+++ tests/playwright/sanity/modules/nested-tabs/nested-tabs-1.test.ts-snapshots/tabs-direction-right-carousel-visible-linux.jpeg
