+++ tests/playwright/sanity/modules/nested-tabs/nested-tabs-2.test.ts-snapshots/tabs-direction-top-icon-position-right-align-left-linux.jpeg
