+++ tests/playwright/sanity/modules/nested-tabs/nested-tabs-4.test.ts-snapshots/tabs-horizontal-scroll-initial-frontend-linux.png
