+++ tests/playwright/sanity/modules/nested-tabs/nested-tabs-4.test.ts-snapshots/tabs-with-hover-mobile-linux.png
